refactor(tool): extract shared import-override helpers into imports.go
resolveAliasOverrides, replaceQualifierAliases, and usedRuleImports lived
inside apply_call.go, scoped to call rules only. Move them into a shared
imports.go alongside a new resolveImportOverrides wrapper, so the other
rule kinds can reuse the same override-rewrite mechanism instead of each
reimplementing it.

Also simplifies FindFuncDecl's rule-type dispatch to a typed field instead
of a bool plus a second type assertion.
